refactor: stream 공통 로직, 타입 변환 등 utils로 따로 분리
