+++ backend/was/src/common/utils/stream.ts
@@ -1,0 +1,23 @@
+export function string2Uint8Array(str: string): Uint8Array {
+  return new TextEncoder().encode(str);
+}
+
+export function uint8Array2String(uint8Array: Uint8Array): string {
+  return new TextDecoder().decode(uint8Array).trim();
+}
+
+export async function string2Uint8ArrayStream(
+  input: string,
+): Promise<ReadableStream<Uint8Array>> {
+  const encoder = new TextEncoder();
+  const uint8Array = encoder.encode(input);
+
+  const readableStream = new ReadableStream({
+    async start(controller) {
+      controller.enqueue(uint8Array);
+      controller.close();
+    },
+  });
+
+  return readableStream;
+}
